+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Linting
 
 - Don't lint pipeline name if `manifest.name` in `.nf-core.yml` ([#2035](https://github.com/nf-core/tools/pull/2035))
+- Don't check for `docker pull` commands in `actions_ci` lint test (leftover from DSL1)
 
 ### General